Rename cluster_setup.py to fixtures.py
