+++ functional_tests/test_single_user_single_channel_doc_updates.py
@@ -3,7 +3,7 @@ import time
 from lib.user import User
 from lib.admin import Admin
 
-from cluster_setup import cluster
+from fixtures import cluster
 
 # Scenario-2:
 # Single User Single Channel: Create Unique docs and update docs verify all num docs present in changes feed.
